Add support to pre-create subports at each trunk
It adds support to precreate subports at each trunk when kuryr is
used, i.e., it allows creating ports and attaching them to the
worker VMs node as subports of their trunk.

It made use of a new variable to decide on creating or not the
subports and how many to create per trunk. The variable is named
openshift_kuryr_precreate_subports. It is false by default.
If included, its value should be equal to the number of subports
to create for each worker VM node.